Added JSON helper and moved some classes
