+++ java/src/main/java/com/awslabs/aws/iot/resultsiterator/helpers/interfaces/JsonHelper.java
@@ -1,0 +1,7 @@
+package com.awslabs.aws.iot.resultsiterator.helpers.interfaces;
+
+public interface JsonHelper {
+    String toJson(Object object);
+
+    <T> T fromJson(Class<T> clazz, byte[] json);
+}
