+++ java/src/main/java/com/awslabs/aws/iot/resultsiterator/helpers/v1/V1HelperModule.java
@@ -1,9 +1,11 @@
 package com.awslabs.aws.iot.resultsiterator.helpers.v1;
 
-import com.awslabs.aws.iot.resultsiterator.helpers.GreengrassIdExtractor;
-import com.awslabs.aws.iot.resultsiterator.helpers.IoHelper;
 import com.awslabs.aws.iot.resultsiterator.helpers.implementations.BasicGreengrassIdExtractor;
 import com.awslabs.aws.iot.resultsiterator.helpers.implementations.BasicIoHelper;
+import com.awslabs.aws.iot.resultsiterator.helpers.implementations.BasicJsonHelper;
+import com.awslabs.aws.iot.resultsiterator.helpers.interfaces.GreengrassIdExtractor;
+import com.awslabs.aws.iot.resultsiterator.helpers.interfaces.IoHelper;
+import com.awslabs.aws.iot.resultsiterator.helpers.interfaces.JsonHelper;
 import com.awslabs.aws.iot.resultsiterator.helpers.v1.implementations.*;
 import com.awslabs.aws.iot.resultsiterator.helpers.v1.interfaces.*;
 import com.google.inject.AbstractModule;
@@ -18,5 +20,6 @@ public class V1HelperModule extends AbstractModule {
         bind(V1ThingHelper.class).to(BasicV1ThingHelper.class);
         bind(V1PolicyHelper.class).to(BasicV1PolicyHelper.class);
         bind(V1ThingGroupHelper.class).to(BasicV1ThingGroupHelper.class);
+        bind(JsonHelper.class).to(BasicJsonHelper.class);
     }
 }
